1: create 3 fragments for FrameLayout
