--- app/src/main/java/ya/co/yandex_finance/MainActivity.kt
@@ -1,44 +1,0 @@
-package ya.co.yandex_finance
-
-import android.os.Bundle
-import android.support.design.widget.BottomNavigationView
-import com.arellomobile.mvp.MvpAppCompatActivity
-import com.arellomobile.mvp.presenter.InjectPresenter
-import kotlinx.android.synthetic.main.activity_main.*
-
-class MainActivity : MvpAppCompatActivity(), BalanceView {
-
-    @InjectPresenter
-    lateinit var balancePresenter: BalancePresenter
-
-    private val mOnNavigationItemSelectedListener = BottomNavigationView.OnNavigationItemSelectedListener { item ->
-        when (item.itemId) {
-            R.id.navigation_home -> {
-                return@OnNavigationItemSelectedListener true
-            }
-            R.id.navigation_dashboard -> {
-                return@OnNavigationItemSelectedListener true
-            }
-            R.id.navigation_notifications -> {
-                return@OnNavigationItemSelectedListener true
-            }
-        }
-        false
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        balancePresenter.updateBalance()
-
-        navigation.setOnNavigationItemSelectedListener(mOnNavigationItemSelectedListener)
-    }
-
-
-    override fun setBalance(balance: Int) {
-        //todo: make main_currency, sub_currency1
-        rub_balance_amount.text = balance.toString()
-//        usd_balance_amount.text =
-    }
-}
