+++ app/src/main/java/ya/co/yandex_finance/component/fragment/AboutFragment.kt
@@ -1,0 +1,20 @@
+package ya.co.yandex_finance.component.fragment
+
+import android.os.Bundle
+import android.support.v4.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import ya.co.yandex_finance.R
+
+
+class AboutFragment : Fragment() {
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
+                              savedInstanceState: Bundle?): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_about, container, false)
+    }
+
+
+}
