+++ app/src/main/res/values/strings.xml
@@ -9,9 +9,14 @@
 
     <!--about screen-->
     <string name="nav_about">О приложении</string>
+    <string name="about_text">Приложение для управления финансами <b><i>Яндекс.Финансы</i></b>.
+        \n\nУправляйте и следите за своим денежным потоком во всех валютах</string>
+
 
     <!--settings screen-->
     <string name="nav_settings">Настройки</string>
     <string name="usd_balance_sign">$</string>
     <string name="rub_balance_sign">₽</string>
+
+
 </resources>
